[RW-2342]	 API: Create API for Preview Data Table (#2036)
* data set preview

* data set preview correct method name

* No of rows a s const

* Fix tests and update logic to work with multiple domains

* Move test helpers up

* Fix spelling error

* Fix codacy

* Add some comments
